Flatten vendored astroq/ttp/ package and fix internal imports
Move the upstream TTP python package up one level (astroq/ttp/ttp/*.py
-> astroq/ttp/*.py) so the import path matches the original ttp.*
namespace. Drop upstream packaging/build artifacts that do not belong
inside AstroQ (bin/, build/, tutorial/, setup.py). Preserve the
upstream README as UPSTREAM_README.md for attribution.

Replace internal unqualified `import star` (formatting.py) and
`import ttp.star as star` (plotting.py) with the explicit
`from astroq.ttp import star`. Replace the upstream sys.path hack in
__init__.py with a docstring crediting the upstream project.
