+++ astroq/ttp/__init__.py
@@ -1,0 +1,5 @@
+"""TTP (Traveling Telescope Problem) solver, vendored into AstroQ.
+
+Upstream: https://github.com/lukehandley/ttp (see UPSTREAM_README.md).
+Reference: Handley et al. 2024, arXiv:2310.18497.
+"""
