+++ astroq/ttp/formatting.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pandas as pd
-import ttp.star as star
+from astroq.ttp import star
 
-def theTTP(filename):
+def theTTP(filename, observatory, nightstarts, nightends):
     """Read in a .csv file of targets and convert them to star objects
 
     Args:
@@ -12,20 +12,21 @@ def theTTP(filename):
         all_star_objects (list): List of star objects, one for each row
             in the .csv file
     """
-
     targets = pd.read_csv(filename)
-
-    if list(targets.columns) != ['Starname', 'RA', 'Dec', 'Exposure Time', 'Exposures Per Visit', 'Visits In Night', 'Intra_Night_Cadence', 'Priority']:
+    required_columns = ['Starname', 'RA', 'Dec', 'Exposure Time', 'Exposures Per Visit', 'Visits In Night', 'Intra_Night_Cadence', 'Priority']
+    optional_columns = ['First Available', 'Last Available']
+    
+    # Check if the first 8 columns match the required columns
+    if list(targets.columns)[:8] != required_columns:
         print(list(targets.columns))
         print("Error: column names not correct.")
-        print("Column names must be this format and this order: ['Starname', 'RA', 'Dec', 'Exposure Time', 'Exposures Per Visit', 'Visits In Night', 'Intra_Night_Cadence', 'Priority'].")
+        print("Column names must be this format and this order: ['Starname', 'RA', 'Dec', 'Exposure Time', 'Exposures Per Visit', 'Visits In Night', 'Intra_Night_Cadence', 'Priority'] or can include ['First Available', 'Last Available'] at the end.")
         return
     else:
         print("Building Star objects:")
         all_star_objects = []
         for t, row in targets.iterrows():
-            starobj = star.star(row, t)
-            # starobj.printStar()
+            starobj = star.star(row, t, observatory, nightstarts, nightends)
             all_star_objects.append(starobj)
         return all_star_objects
