+++ astroq/ttp/model.py
