--- astroq/ttp/ttp/__init__.py
@@ -1,5 +1,0 @@
-import os
-import sys
-
-_ROOT = os.path.abspath(os.path.dirname(__file__))
-sys.path.append(_ROOT)
