--- astroq/ttp/tutorial/TKStargets.csv
@@ -1,87 +1,0 @@
-Starname,RA,Dec,Exposure Time,Exposures Per Visit,Visits In Night,Intra_Night_Cadence,Priority
-TOI-260,4.7732,-9.9648,644,1,1,0,10
-TOI-266,26.2096,-18.4009,143,1,1,0,10
-TOI-329,351.3209,-15.6347,403,1,1,0,10
-TOI-465,32.7818,2.4180,416,1,1,0,10
-TOI-469,93.0582,-14.6500,663,1,1,0,10
-TOI-480,88.3787,-16.2650,414,1,1,0,10
-TOI-509,117.9250,9.3861,601,1,2,2,10
-TOI-554,60.7479,9.2085,1200,1,1,0,10
-TOI-561,148.1856,6.2164,1000,1,1,0,10
-TOI-669,158.9006,-5.1817,248,1,1,0,10
-TOI-1136,192.1849,64.8553,37,1,1,0,10
-TOI-1173,197.6823,70.7684,639,1,1,0,10
-TOI-1174,209.2181,68.6180,117,1,1,0,10
-TOI-1180,214.5531,82.1940,42,1,1,0,10
-TOI-1181,297.2159,64.3544,104,1,1,0,10
-TOI-1184,272.2039,60.6782,658,1,1,0,10
-TOI-1194,167.8205,69.9647,1200,1,1,0,10
-TOI-1244,256.2803,69.5194,399,1,1,0,10
-TOI-1246,251.1165,70.4296,888,1,1,0,10
-TOI-1247,227.8705,71.8410,590,1,1,0,10
-TOI-1248,259.0236,63.1060,199,1,1,0,10
-TOI-1249,200.5617,66.3087,71,1,1,0,10
-TOI-1255,296.2443,74.0627,701,1,1,0,10
-TOI-1269,249.6971,64.5587,828,1,1,0,10
-TOI-1272,199.1966,49.8610,205,1,1,0,10
-TOI-1279,185.0639,56.2011,346,1,1,0,10
-TOI-1288,313.1666,65.6091,64,1,1,0,10
-TOI-1294,223.0929,70.4766,327,1,1,0,10
-TOI-1296,256.7709,70.2385,767,1,1,0,10
-TOI-1298,241.3234,70.1899,575,1,1,0,10
-TOI-1339,302.0240,66.8506,31,1,1,0,10
-TOI-1347,280.3268,70.2899,794,1,1,0,10
-TOI-1386,334.5037,54.3190,478,1,1,0,10
-TOI-1410,334.8832,42.5603,623,1,1,0,10
-TOI-1411,232.9451,47.0568,203,1,1,0,10
-TOI-1422,354.2412,39.6394,740,1,1,0,10
-TOI-1430,300.6143,53.3768,1000,1,1,0,10
-TOI-1436,215.6025,55.3337,304,1,1,0,10
-TOI-1437,256.1387,56.8426,604,1,1,0,10
-TOI-1438,280.9243,74.9376,402,1,1,0,10
-TOI-1439,241.7639,67.8777,463,1,1,0,10
-TOI-1443,297.4003,76.1391,177,1,1,0,10
-TOI-1444,305.4743,70.9438,757,1,1,0,10
-TOI-1451,186.5243,61.2590,209,1,1,0,10
-TOI-1456,306.7413,33.7445,163,1,1,0,10
-TOI-1467,19.1139,49.2338,234,1,1,0,10
-TOI-1471,30.9043,21.2809,468,1,1,0,10
-TOI-1472,14.1136,48.6376,100,1,1,0,10
-TOI-1473,15.5982,37.1855,729,1,1,0,10
-TOI-1601,38.3614,41.0134,659,1,1,0,10
-TOI-1611,325.1866,84.3335,222,1,1,0,10
-TOI-1669,45.9529,83.5876,559,1,1,0,10
-TOI-1691,272.4061,86.8596,660,1,1,0,10
-TOI-1694,97.7482,66.3607,549,1,1,0,10
-TOI-1710,94.2828,76.2108,711,1,1,0,10
-TOI-1716,105.0829,56.8244,503,1,1,0,10
-TOI-1723,116.7971,68.4766,184,1,1,0,10
-TOI-1726,117.4794,27.3632,536,1,1,0,10
-TOI-1736,43.4350,69.1014,380,1,1,0,10
-TOI-1742,257.3285,71.8764,248,1,1,0,10
-TOI-1751,243.4888,63.5343,111,1,1,0,10
-TOI-1753,252.4698,61.1735,257,1,1,0,10
-TOI-1758,354.7430,75.6851,564,1,1,0,10
-TOI-1759,326.8533,62.7539,743,1,1,0,10
-TOI-1775,150.1151,39.4578,302,1,1,0,10
-TOI-1776,164.7761,40.9836,1200,1,1,0,10
-TOI-1778,136.7781,46.6726,638,1,1,0,10
-TOI-1794,203.3977,49.0611,616,1,1,0,10
-TOI-1797,162.7771,25.6412,693,1,1,0,10
-TOI-1798,211.0941,46.5194,331,1,1,0,10
-TOI-1799,167.2330,34.3032,634,1,1,0,10
-TOI-1801,175.5766,23.0269,553,1,1,0,10
-TOI-1807,201.2833,38.9225,763,1,1,0,10
-TOI-1823,196.2204,63.7538,440,1,1,0,10
-TOI-1824,197.7312,61.7448,424,1,1,0,10
-TOI-1836,245.9082,54.6898,232,1,1,0,10
-TOI-1842,201.9628,9.0307,314,1,1,0,10
-TOI-1898,144.5556,23.5469,213,1,1,0,10
-TOI-1905,188.3869,-10.1461,462,1,1,0,10
-TOI-2019,234.4317,48.9554,107,1,1,0,10
-TOI-2045,1.1191,54.9345,571,1,1,0,10
-TOI-2076,217.3927,39.7904,146,1,1,0,10
-TOI-2088,261.3752,75.8823,345,1,1,0,10
-TOI-2114,261.0964,33.2051,199,1,1,0,10
-TOI-2128,256.9826,32.1055,362,1,1,0,10
-TOI-2145,263.7581,40.6951,398,1,1,0,10
