--- astroq/ttp/tutorial/inputs.txt
@@ -1,5 +1,0 @@
-StarListFile : ../tutorial/TKStargets.csv
-StartNight : 2024-02-11T07:00:00
-EndNight : 2024-02-11T14:00:00
-Telescope : Keck1
-SavePath : ../tutorial/outputs/
