--- astroq/ttp/tutorial/outputs/ObserveOrder_2024-02-11.txt
@@ -1,53 +1,0 @@
-Target,StartExposure
-TOI-465,07:00
-TOI-1898,07:08
-TOI-1473,07:13
-TOI-1472,07:26
-TOI-1467,07:27
-TOI-1471,07:31
-TOI-1601,07:40
-TOI-1669,07:51
-TOI-1736,08:00
-TOI-480,08:09
-TOI-469,08:16
-TOI-509,08:28
-TOI-669,08:39
-TOI-1775,08:44
-TOI-1799,08:49
-TOI-1776,09:00
-TOI-1710,09:20
-TOI-554,09:33
-TOI-1694,09:54
-TOI-1723,10:03
-TOI-1778,10:06
-TOI-1797,10:18
-TOI-1279,10:30
-TOI-1272,10:36
-TOI-1794,10:39
-TOI-1807,10:50
-TOI-1842,11:03
-TOI-1905,11:08
-TOI-1801,11:17
-TOI-2076,11:26
-TOI-1798,11:29
-TOI-1436,11:34
-TOI-1824,11:47
-TOI-1823,11:54
-TOI-1716,12:01
-TOI-1726,12:10
-TOI-1194,12:20
-TOI-1451,12:40
-TOI-1136,12:44
-TOI-509,12:47
-TOI-1411,12:59
-TOI-2019,13:02
-TOI-1691,13:05
-TOI-1180,13:16
-TOI-1173,13:16
-TOI-1249,13:27
-TOI-1174,13:28
-TOI-1294,13:30
-TOI-1247,13:36
-TOI-1836,13:46
-TOI-1751,13:50
-TOI-1439,13:52
